% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pgnmergen613-package.R
\docType{package}
\name{hw04pgnmergen613-package}
\alias{hw04pgnmergen613}
\alias{hw04pgnmergen613-package}
\title{hw04pgnmergen613: What the Package Does (One Line, Title Case)}
\description{
Finds nth element of a recursive sequence defined by a formula,
              which is then plotted by a linegraph.
}
\author{
\strong{Maintainer}: Mergentevne Narangerel \email{mergen.fortuna@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
