% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Recursive Sequence Element Find}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{vector consisting of three positive numbers.}

\item{n}{number for the nth element of the sequence.}
}
\value{
X value of the nth element using the sequence.
}
\description{
Takes a positive numeric vector of 3 and finds the nth element of the
sequence following the formula: x{n} = x{n-1} + (x{n-3}-x{n-2})/4.
}
\examples{
myseq_n(x = c(2,4,3), n = 5)
}
