% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Plotter}
\usage{
plotter(c.table)
}
\arguments{
\item{c.table}{Dataframe of 4 columns consisting of positive numbers.}
}
\value{
Uses myseq_n function to plot the sequence of numbers of nth element.
}
\description{
User submitted numeric dataframe of 4 columns is plotted as a line graph.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
}
